+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.4.3 (pre-release)
 * [Fix #158] NRE when transferring crew into a pod part without KIS inventories.
 * [Fix #199] Wrong kerbal class in inventory window info when in command seat.
+* [Fix #203] Some physicsless part trigger physics effect on move&attach.
 * [Enhancement #200] Set decouple momentum on Cargo Mount to zero.
 * [Enhancement #202] Make pod seat's inventory more clear.